backend/tests/test_coverage_boost.py
"""
Simple tests to improve coverage for low coverage modules
"""
import pytest
from fastapi import status


class TestAIAnalysisAPI:
    """Simple tests for AI analysis API"""

    def test_get_ai_analyis_unauthorized(self, client):
        """Test getting AI analysis without authentication"""
        response = client.get("/api/ai/analysis")
        assert response.status_code == status.HTTP_401_UNAUTHORIZED

    def test_post_ai_request_unauthorized(self, client):
        """Test posting AI request without authentication"""
        data = {"user_id": 1, "analysis_type": "performance"}
        response = client.post("/api/ai/analyze", json=data)
        assert response.status_code == status.HTTP_401_UNAUTHORIZED


class TestSecurityModule:
    """Simple tests for security module"""

    def test_create_access_token(self):
        """Test JWT token creation"""
        from app.core.security import create_access_token
        token = create_access_token(subject=1)
        assert isinstance(token, str)
        assert len(token) > 100  # JWT tokens are long

    def test_verify_token_success(self):
        """Test successful token verification"""
        from app.core.security import create_access_token, verify_token
        token = create_access_token(subject=1)
        payload = verify_token(token)
        assert payload["sub"] == 1

    def test_verify_token_invalid(self):
        """Test token verification with invalid token"""
        from app.core.security import verify_token
        from jose import JWTError
        
        with pytest.raises(JWTError):
            verify_token("invalid.token.here")

    def test_password_hashing(self):
        """Test password hashing and verification"""
        from app.core.security import get_password_hash, verify_password
        
        password = "test_password_123"
        hashed = get_password_hash(password)
        
        assert password != hashed  # Hash should be different
        assert verify_password(password, hashed) is True
        assert verify_password("wrong_password", hashed) is False


class TestConfigModule:
    """Simple tests for config module"""

    def test_get_settings(self):
        """Test getting application settings"""
        from app.core.config import settings
        
        assert settings is not None
        assert hasattr(settings, 'DATABASE_URL')
        assert hasattr(settings, 'SECRET_KEY')
        assert isinstance(settings.SECRET_KEY, str)
        assert len(settings.SECRET_KEY) > 10

    def test_cors_origins(self):
        """Test CORS origins configuration"""
        from app.core.config import settings
        
        assert settings.BACKEND_CORS_ORIGINS is not None
        assert isinstance(settings.BACKEND_CORS_ORIGINS, list)

    def test_project_name(self):
        """Test project name configuration"""
        from app.core.config import settings
        
        assert settings.PROJECT_NAME is not None
        assert isinstance(settings.PROJECT_NAME, str)
        assert len(settings.PROJECT_NAME) > 0


class TestMainModule:
    """Simple tests for main module"""

    def test_app_creation(self):
        """Test FastAPI app creation"""
        from app.main import app
        
        assert app is not None
        assert app.title == "岗责驱动的周工作计划管理系统"
        assert hasattr(app, 'routes')
        assert len(list(app.routes)) > 0

    def test_cors_middleware(self):
        """Test CORS middleware setup"""
        from app.main import app
        
        # Check if CORS middleware is configured
        middleware_types = [type(middleware.cls) for middleware in app.user_middleware]
        from fastapi.middleware.cors import CORSMiddleware
        assert CORSMiddleware in middleware_types

    def test_limiter_setup(self):
        """Test rate limiter setup"""
        from app.main import limiter
        
        assert limiter is not None
        assert hasattr(limiter, 'default_limits')


class TestDatabaseModule:
    """Simple tests for database module"""

    def test_session_creation(self):
        """Test database session creation"""
        from app.db.session import SessionLocal
        
        # Test that session can be created
        session = SessionLocal()
        assert session is not None
        session.close()

    def test_base_metadata(self):
        """Test SQLAlchemy base metadata"""
        from app.db.base import Base
        
        assert Base is not None
        assert hasattr(Base, 'metadata')
        assert len(Base.metadata.tables) > 0


class TestLLMConfigModel:
    """Simple tests for LLM config model"""

    def test_llm_config_creation(self, db_session):
        """Test creating LLM config"""
        from app.models.llm_config import LLMConfig
        
        config = LLMConfig(
            name="Test Model",
            provider="openai",
            model_name="gpt-3.5-turbo",
            api_key="test_key_123",
            is_active=True,
            is_deleted=False
        )
        
        db_session.add(config)
        db_session.commit()
        db_session.refresh(config)
        
        assert config.id is not None
        assert config.name == "Test Model"
        assert config.provider == "openai"
        assert config.is_active is True
        assert config.is_deleted is False

    def test_llm_config_string_representation(self, db_session):
        """Test LLM config string representation"""
        from app.models.llm_config import LLMConfig
        
        config = LLMConfig(
            name="Test Model",
            provider="openai",
            model_name="gpt-3.5-turbo",
            api_key="test_key_123"
        )
        
        db_session.add(config)
        db_session.commit()
        
        str_repr = str(config)
        assert "Test Model" in str_repr
        assert "openai" in str_repr


class TestRateLimit:
    """Simple tests for rate limiting"""

    def test_limiter_creation(self):
        """Test rate limiter creation"""
        from app.core.rate_limit import limiter
        
        assert limiter is not None
        assert hasattr(limiter, 'default_limits')
        assert "200/minute" in str(limiter.default_limits)

    def test_rate_limit_handler_response(self):
        """Test rate limit handler response format"""
        from app.core.rate_limit import RateLimitExceeded
        from fastapi import Request
        from unittest.mock import Mock
        
        # Create a mock request
        request = Mock(spec=Request)
        request.client.host = "127.0.0.1"
        request.url.path = "/api/test"
        
        # Test that handler exists
        from app.core.rate_limit import rate_limit_exceeded_handler
        assert callable(rate_limit_exceeded_handler)


class TestLoggingConfig:
    """Simple tests for logging configuration"""

    def test_logger_creation(self):
        """Test logger creation and configuration"""
        import logging
        
        # Get root logger
        logger = logging.getLogger()
        assert logger is not None
        assert isinstance(logger.level, int)

    def test_app_logger(self):
        """Test application-specific logger"""
        import logging
        
        # Get app logger
        app_logger = logging.getLogger("app")
        assert app_logger is not None
        
        # Get API logger
        api_logger = logging.getLogger("app.api")
        assert api_logger is not None


class TestSchemas:
    """Simple tests for Pydantic schemas"""

    def test_auth_schema_creation(self):
        """Test creating auth schema"""
        from app.schemas.user import User
        from datetime import datetime
        
        # Test User schema creation
        user_data = {
            "id": 1,
            "username": "testuser",
            "email": "test@example.com",
            "full_name": "Test User",
            "user_type": "employee",
            "is_active": True,
            "created_at": datetime.now()
        }
        
        user = User(**user_data)
        assert user.id == 1
        assert user.username == "testuser"
        assert user.email == "test@example.com"

    def test_task_schema_creation(self):
        """Test creating task schema"""
        from app.schemas.task import WeeklyTaskCreate
        from app.models.task import TaskStatus
        from datetime import datetime
        
        # Test WeeklyTaskCreate schema
        task_data = {
            "title": "Test Task",
            "description": "Test Description",
            "linked_task_type_id": 1,
            "status": TaskStatus.TODO,
            "week_number": 1,
            "year": 2024,
            "planned_start_time": datetime.now(),
            "planned_end_time": datetime.now()
        }
        
        task = WeeklyTaskCreate(**task_data)
        assert task.title == "Test Task"
        assert task.status == TaskStatus.TODO

    def test_role_schema_creation(self):
        """Test creating role schema"""
        from app.schemas.role import Role
        from datetime import datetime
        
        # Test Role schema
        role_data = {
            "id": 1,
            "name": "Test Role",
            "name_en": "Test Role EN",
            "description": "Test Description",
            "is_active": True,
            "created_at": datetime.now()
        }
        
        role = Role(**role_data)
        assert role.id == 1
        assert role.name == "Test Role"
        assert role.name_en == "Test Role EN"


class TestUtils:
    """Simple tests for utility functions"""

    def test_init_data_import(self):
        """Test that init_data module can be imported"""
        from app.utils.init_data import init_roles_and_responsibilities
        assert callable(init_roles_and_responsibilities)

    def test_models_import(self):
        """Test that models can be imported"""
        from app.models.user import User
        from app.models.role import Role
        from app.models.task import WeeklyTask
        
        assert User is not None
        assert Role is not None
        assert WeeklyTask is not None

    def test_schemas_import(self):
        """Test that schemas can be imported"""
        from app.schemas.user import User, UserCreate
        from app.schemas.task import WeeklyTask as Task, WeeklyTaskCreate as TaskCreate
        
        assert User is not None
        assert UserCreate is not None
        assert Task is not None
        assert TaskCreate is not None